Warnings

+++ src/drivers/mc/mc_init.c
@@ -70,7 +70,7 @@ static int dpci_discovery()
 	struct dprc_obj_desc dev_desc;
 	int err = 0;
 	int i = 0;
-	int p = 0;;
+	uint8_t p = 0;;
 	struct dprc *dprc = sys_get_unique_handle(FSL_OS_MOD_AIOP_RC);
 	struct dpci *dpci;
 	struct dpci_dest_cfg dest_cfg;
